Updated media queries
added media queries to make the spacing better in different sections.

on the hero section the background image was not covering the whole section. It has now been updated.

the text on hero and about was being pushed down, on big screens (780px or more). It has now been fixed.

next steps would be to add hover effects  on buttons and links.

Also, add a preview  pictures on Figma links. :)